requirements.txt
config==0.3.9
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0

djangorestframework==3.9.4
django-bootstrap3==11.0.0
django-heroku==0.3.1
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.7.5
python-decouple==3.1
pytz==2019.1
whitenoise==4.1.2
